refactor: Move process_skill function to packager module
- Relocate process_skill function from agents.py to packager.py
- Update import statements and type hints
- Maintain existing function logic and error handling
- Improve code organization by keeping skill processing logic in dedicated module
